# backend/apidocs.yaml
openapi: 3.0.0
info:
  title: IoT Project API
  version: 1.0.0
  description: API for IoT Project managing sensor data and device actions
servers:
  - url: http://localhost:8081
paths:
  /status:
    get:
      summary: Get latest status data
      description: Retrieves the latest status data for fan, light, AC, temperature, humidity, and lux
      responses:
        '200':
          description: Successful response
          content:
            application/json:    
              schema:
                type: array
                items:
                  type: object
                  properties:
                    fan: 
                      type: boolean
                    light: 
                      type: boolean
                    ac: 
                      type: boolean
                    temperature: 
                      type: number
                    humidity: 
                      type: number
                    lux: 
                      type: number
        '500':
          description: Server error
  /data1:
    get:
      summary: Get all sensor data
      description: Retrieves all sensor data from the database
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    temperature:
                      type: number
                    humidity:
                      type: number
                    lux:
                      type: number
                    date:
                      type: string
                      format: date-time
        '500':
          description: Server error
  /data2:
    get:
      summary: Get all action history
      description: Retrieves all action history data from the database
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    device:
                      type: string
                    action:
                      type: string
                    date:
                      type: string
                      format: date-time
        '500':
          description: Server error
  /actiondata:
    post:
      summary: Send action data
      description: Sends action data to control a device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device:
                  type: string
                action:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  device:
                    type: string
                  action:
                    type: string
        '500':
          description: Server error
  /datasearch1:
    get:
      summary: Search sensor data
      description: Search and filter sensor data
      parameters:
        - in: query
          name: searchTerm
          schema:
            type: string
        - in: query
          name: dateFilter
          schema:
            type: string
        - in: query
          name: parameterFilter
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    temperature:
                      type: number
                    humidity:
                      type: number
                    lux:
                      type: number
                    date:
                      type: string
                      format: date-time
        '500':
          description: Server error
  /datasearch2:
    get:
      summary: Search action history
      description: Search and filter action history data
      parameters:
        - in: query
          name: searchTerm
          schema:
            type: string
        - in: query
          name: dateFilter
          schema:
            type: string
        - in: query
          name: deviceFilter
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    device:
                      type: string
                    action:
                      type: string
                    date:
                      type: string
                      format: date-time
        '500':
          description: Server error
  /sort1:
    get:
      summary: Sort sensor data
      description: Sort sensor data by specified column and order
      parameters:
        - in: query
          name: column
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
        - in: query
          name: parameterFilter
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    temperature:
                      type: number
                    humidity:
                      type: number
                    lux:
                      type: number
                    date:
                      type: string
                      format: date-time
        '500':
          description: Server error
  /sort2:
    get:
      summary: Sort action history
      description: Sort action history data by specified column and order
      parameters:
        - in: query
          name: column
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    device:
                      type: string
                    action:
                      type: string
                    date:
                      type: string
                      format: date-time
        '500':
          description: Server error